ozone-framework-python-server/requirements.txt
alabaster==0.7.12
autopep8==1.4.4
Babel==2.7.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.1
django-filter==2.1.0
django-rest-swagger==2.2.0
djangorestframework==3.9.4
drf-extensions==0.5.0
idna==2.8
imagesize==1.1.0
itypes==1.1.0
Jinja2==2.10.1
Markdown==3.1.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==19.0
psycopg2==2.8.2
pycodestyle==2.5.0
Pygments==2.4.2
pyparsing==2.4.1.1
pytz==2019.1
requests==2.22.0
simplejson==3.16.0
six==1.12.0
snowballstemmer==1.9.0
Sphinx==2.1.2
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
uuid==1.30
